// client/src/common/models/ViewTypes.ts
export type OverridableOptions = {
  keyColour?: string;
  textColour?: string;
  textBackground?: string;
  font?: string;
  size?: number;
  justifyContent?: 'start' | 'center' | 'end';
  alignItems?: 'start' | 'center' | 'end';
  left?: string;
  top?: string;
  hideNav?: boolean;
  hideOvertime?: boolean;
  hideMessagesOverlay?: boolean;
}